requirements.txt
anyio==4.10.0
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
python-dotenv==1.0.0
python-telegram-bot==20.7
